🔨 Refactor: Header, Footer, Modal 컴포넌트 index.js로 엔트리 변경
컴포넌트 import 경로를 간결하게 하기 위해 Header, Footer, Modal 디렉토리에 index.js 추가
구조 정리에 따른 재사용성과 가독성 개선 목적

Resolves: #34
